final_project_comp/requirements.txt
numpy>=1.26.4
pandas>=2.2.1
scikit-learn>=1.4.1
fastapi>=0.110.0
uvicorn>=0.27.1
pytest>=8.0.2
joblib>=1.3.2
pydantic>=2.6.3
python-dotenv>=1.0.1
matplotlib>=3.8.3
seaborn>=0.13.2